Move micromark.cjs helpers from lib to helpers so it can be used by code in both locations.
