+++ lib/markdownlint.js
@@ -5,7 +5,7 @@
 const path = require("node:path");
 const { promisify } = require("node:util");
 const markdownit = require("markdown-it");
-const micromark = require("./micromark.cjs");
+const micromark = require("../helpers/micromark.cjs");
 const { deprecatedRuleNames } = require("./constants");
 const rules = require("./rules");
 const helpers = require("../helpers");
